files/folders renamed. pom.xml changes reverted. fixed api base url.
