+++ src/main/js/app/api/AssetApi.js
@@ -2,12 +2,12 @@ import axios from 'axios';
 import config from './config';
 
 const axiosInstance = axios.create({
+    baseURL: config.baseUrl,
     timeout: 10000,
     withCredentials: false,
     headers: {
         Accept: 'application/json',
         'Content-Type': 'application/json',
-        'Access-Control-Allow-Origin': '*'
     }
 });
