+++ src/main/js/app/api/LoginApi.js
@@ -8,7 +8,6 @@ const axiosInstance = axios.create({
     headers: {
         Accept: 'application/json',
         'Content-Type': 'application/json',
-        'Access-Control-Allow-Origin': '*'
     }
 });
